+++ src/component/page/edit/Node/CardNode/index.tsx
@@ -1,10 +1,10 @@
 import { Flex, ListItem } from '@chakra-ui/react';
 import { FC } from 'react';
 
-import { CardTag } from '@/page/edit/Tag/CardTag';
 import { CardNodeType } from '@/types/schema';
 
 import { PositionNode } from '../PositionNode';
+import { CardTag } from '../../Tag/CardTag';
 
 export const CardNode: FC<CardNodeType & { path: Array<number | string> }> = (cardNode) => {
   return (
