+++ src/pages/edit/[id].tsx
@@ -5,8 +5,8 @@ import { useRouter } from 'next/router';
 import nookies from 'nookies';
 
 import { MobilePage } from '@/component/layout/MobilePage';
+import { PostPage } from '@/component/page/edit/PostPage';
 import { adminApp } from '@/lib/firebase/init/server';
-import { PostPage } from '@/page/edit/PostPage';
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {
   try {
