+++ src/pages/edit/index.tsx
@@ -4,8 +4,8 @@ import { GetServerSideProps } from 'next';
 import nookies from 'nookies';
 
 import { MobilePage } from '@/component/layout/MobilePage';
+import { HandTreePage } from '@/component/page/edit';
 import { adminApp } from '@/lib/firebase/init/server';
-import { HandTreePage } from '@/page/edit';
 
 export const getServerSideProps: GetServerSideProps = async (ctx) => {
   try {
